% Generated by roxygen2 (4.0.2): do not edit by hand
\name{metaData}
\alias{metaData}
\title{Retrieve metadata from a Quandl series}
\usage{
metaData(x)
}
\arguments{
\item{x}{A Quandl time series object with attached meta data.}
}
\value{
Returns a list of meta data about the series.
}
\description{
Retrieve metadata from a Quandl series
}
\examples{
\dontrun{
metaData(ts)
}
}
\seealso{
\code{\link{Quandl}}
}

