% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Quandlconfig.R
\name{Quandl.api_key}
\alias{Quandl.api_key}
\title{Query or set Quandl API key}
\usage{
Quandl.api_key(api_key)
}
\arguments{
\item{api_key}{Optionally passed parameter to set Quandl \code{api_key}.}
}
\value{
Returns invisibly the currently set \code{api_key}.
}
\description{
Query or set Quandl API key
}
\examples{
\dontrun{
Quandl.api_key('foobar')
}
}

