% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Quandlconfig.R
\name{Quandl.api_version}
\alias{Quandl.api_version}
\title{Query or set Quandl API version}
\usage{
Quandl.api_version(api_version)
}
\arguments{
\item{api_version}{Optionally passed parameter to set Quandl \code{api_version}.}
}
\value{
Returns invisibly the currently set \code{api_version}.
}
\description{
Query or set Quandl API version
}
\examples{
\dontrun{
Quandl.api_version('2015-04-09')
}
}
\seealso{
\code{\link{Quandl}}
}

