% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Quandlconfig.R
\name{Quandl.auth}
\alias{Quandl.auth}
\title{Query or set Quandl API token}
\usage{
Quandl.auth(auth_token)
}
\arguments{
\item{auth_token}{Optionally passed parameter to set Quandl \code{auth_token}.}
}
\value{
Returns invisibly the currently set \code{auth_token}.
}
\description{
Deprecated. Alias of \code{\link{Quandl.api_key}}
}
\examples{
\dontrun{
Quandl.auth('foobar')
}
}
\seealso{
\code{\link{Quandl}}
}

