% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quandlconfig.R
\name{Quandl.base_url}
\alias{Quandl.base_url}
\title{Query or set Base URL for White-labeled sites}
\usage{
Quandl.base_url(base_url)
}
\arguments{
\item{base_url}{Optionally passed parameter to set Quandl \code{base_url}.}
}
\value{
Returns invisibly the currently set \code{base_url}.
}
\description{
Query or set Base URL for White-labeled sites
}
\examples{
\dontrun{
Quandl.base_url('http://localhost')
}
}
