% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Quandldatabase.R
\name{Quandl.database.download_url}
\alias{Quandl.database.download_url}
\title{Returns the bulk download url}
\usage{
Quandl.database.download_url(database_code, ...)
}
\arguments{
\item{database_code}{Database code on Quandl specified as a string}

\item{...}{Additional named values that are interpretted as api parameters e.g., download_type.}
}
\value{
Returns the download url
}
\description{
An authentication token is needed for access to the Quandl API multiple times. Set your \code{access_token} with \code{Quandl.api_key} function.
}
\details{
For instructions on finding your authentication token go to https://www.quandl.com/account
}
\examples{
\dontrun{
quandldata = Quandl.database.download_url("NSE", download_type="partial")
}
}
\author{
Raymond McTaggart
}
\references{
This R package uses the Quandl API. For more information go to https://www.quandl.com/help/api. For more help on the package itself go to http://www.quandl.com/help/r.
}
\seealso{
\code{\link{Quandl.api_key}}
}

