% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Quandl.R
\name{Quandl}
\alias{Quandl}
\title{Pulls Data from the Quandl Dataset endpoint and formats}
\usage{
Quandl(code, type = c("raw", "ts", "zoo", "xts", "timeSeries"),
  transform = c("", "diff", "rdiff", "normalize", "cumul", "rdiff_from"),
  collapse = c("", "daily", "weekly", "monthly", "quarterly", "annual"),
  order = c("desc", "asc"), meta = FALSE, ...)
}
\arguments{
\item{code}{Dataset code on Quandl specified as a string or an array of strings.}

\item{type}{Type of data returned specified as string. Can be 'raw', 'ts', 'zoo', 'xts' or 'timeSeries'.}

\item{transform}{Apply Quandl API data transformations.}

\item{collapse}{Collapse frequency of Data.}

\item{order}{Select if data is given to R in ascending or descending formats. Helpful for the rows parameter.}

\item{meta}{Returns meta data in list format as well as data.}

\item{...}{Additional named values that are interpreted as api parameters.}
}
\value{
Depending on the type the class is either data.frame, time series, xts, zoo or timeSeries.
}
\description{
An authentication token is needed for access to the Quandl API multiple times. Set your \code{access_token} with \code{Quandl.api_key} function.
}
\details{
For instructions on finding your authentication token go to https://www.quandl.com/account/api
}
\examples{
\dontrun{
quandldata = Quandl("NSE/OIL", collapse="monthly", start_date="2013-01-01", type="ts")
plot(quandldata[,1])
}
}
\author{
Raymond McTaggart
}
\references{
This R package uses the Quandl API. For more information go to https://www.quandl.com/docs/api. For more help on the package itself go to http://www.quandl.com/help/r.
}
\seealso{
\code{\link{Quandl.api_key}}
}

