% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Quandlsearch.R
\name{Quandl.search}
\alias{Quandl.search}
\title{Search the Quandl database}
\usage{
Quandl.search(query, silent = FALSE, per_page = 10, ...)
}
\arguments{
\item{query}{Search terms}

\item{silent}{Prints the first few results when FALSE.}

\item{per_page}{Number of results returned per page.}

\item{...}{Additional named values that are interpretted as api parameters e.g., page.}
}
\value{
A list of the search results.
}
\description{
An authentication token is needed for access to the Quandl API multiple times. Set your \code{access_token} with \code{Quandl.api_key} function.
}
\details{
For instructions on finding your authentication token go to https://www.quandl.com/account/api
}
\examples{
\dontrun{
search.results <- Quandl.search("oil")
}
}
\references{
This R package uses the Quandl API. For more information go to \url{https://www.quandl.com/docs/api}. For more help on the package itself go to \url{http://www.quandl.com/help/r}.
}
\seealso{
\code{\link{Quandl.api_key}}
}

