\name{Quandlr}
\alias{Quandlr}
\docType{data}
\title{Authentication Token Holder}
\description{Environment in which to store the users authentication token}
\usage{data(Quandlr)}
\format{
  The format is:
<environment: 0x9597718> 
}
\details{An authentication token is needed for access to the Quandl API multiple times. This stores it away from the user workspace.}
\references{For instructions on finding your authentication token go to www.quandl.com/API.}
\keyword{datasets}
