% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Quandl.R
\name{metaData}
\alias{metaData}
\title{Retrieve metadata from a Quandl series or search results}
\usage{
metaData(x)
}
\arguments{
\item{x}{A Quandl time series object or search results with attached meta data.}
}
\value{
Returns a list of meta data about the series or search results.
}
\description{
Retrieve metadata from a Quandl series or search results
}
\examples{
\dontrun{
metaData(ts)
}
}
\seealso{
\code{\link{Quandl}}, \code{\link{Quandl.search}}
}

