% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Quandl.curlopts}
\alias{Quandl.curlopts}
\title{Set curl options on all Quandl api calls. e.g. for use with a proxy}
\usage{
Quandl.curlopts(curl)
}
\arguments{
\item{curl}{A curl handle object}
}
\value{
Returns invisibly the currently set \code{curl}
}
\description{
Set curl options on all Quandl api calls. e.g. for use with a proxy
}
\examples{
\dontrun{
Quandl.curlopts(getCurlHandle())
}
}
\seealso{
\code{\link{Quandl}}
}

