% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Quandldatabase.R
\name{Quandl.database.bulk_download_url}
\alias{Quandl.database.bulk_download_url}
\title{Returns the bulk download url}
\usage{
Quandl.database.bulk_download_url(database_code, ...)
}
\arguments{
\item{database_code}{Database code on Quandl specified as a string.}

\item{...}{Additional named values that are interpreted as Quandl API parameters. Please see \url{https://www.quandl.com/docs/api#entire-database} for a full list of parameters.}
}
\value{
Returns the download url.
}
\description{
Returns the bulk download url
}
\details{
Set your \code{api_key} with \code{Quandl.api_key} function. For instructions on finding your api key go to \url{https://www.quandl.com/account/api}
}
\examples{
\dontrun{
url = Quandl.database.download_url("NSE", download_type="partial")
}
}
\seealso{
\code{\link{Quandl.api_key}}
}

